[Feat] 원거리 공격 및 관통 구현

+++ Assets/Workspace/HSD/Scripts/Unit/UnitStatusController.cs
@@ -28,7 +28,7 @@ public class UnitStatusController : MonoBehaviour, IDamageable
     public Stat<int> MagicDefense;
 
     [Header("Range")]
-    public Stat<int> AttackRange;
+    public Stat<float> AttackRange;
     public Stat<int> AttackCount;
     public float DetectionRange = 5f;
     #endregion
